(PC-17806) refactor(auth): rename support.service > contactSupport and move in helpers folder
