+++ jest/jest.setup.ts
@@ -32,7 +32,7 @@ jest.mock('react-native-safe-area-context', () => ({
   useSafeAreaInsets: () => ({ bottom: 16, right: 16, left: 16, top: 16 }),
 }))
 
-jest.mock('features/auth/support.services')
+jest.mock('features/auth/helpers/contactSupport')
 
 /* See the corresponding mocks in features/navigation/RootNavigator/__mocks__ */
 jest.mock('features/navigation/RootNavigator/routes')
