+++ src/features/auth/signup/AcceptCgu/AcceptCgu.tsx
@@ -3,8 +3,8 @@ import { v4 as uuidv4 } from 'uuid'
 
 import { CardContent, Paragraphe } from 'features/auth/components/signupComponents'
 import { useSettingsContext } from 'features/auth/context/SettingsContext'
+import { contactSupport } from 'features/auth/helpers/contactSupport'
 import { PreValidationSignupStepProps } from 'features/auth/signup/types'
-import { contactSupport } from 'features/auth/support.services'
 import { env } from 'libs/environment'
 import { captureMonitoringError } from 'libs/monitoring'
 import { useNetInfoContext } from 'libs/network/NetInfoWrapper'
