+++ src/features/identityCheck/pages/identification/errors/hooks/useNotEligibleEduConnectErrorData.ts
@@ -2,8 +2,8 @@ import { useNavigation } from '@react-navigation/native'
 import { FunctionComponent } from 'react'
 import { TextStyle } from 'react-native'
 
+import { contactSupport } from 'features/auth/helpers/contactSupport'
 import { useBeneficiaryValidationNavigation } from 'features/auth/signup/useBeneficiaryValidationNavigation'
-import { contactSupport } from 'features/auth/support.services'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { ExternalNavigationProps, InternalNavigationProps } from 'ui/components/touchableLink/types'
 import { UserErrorWhite } from 'ui/svg/BicolorUserError'
